use float16 for mochi

+++ tests/context_parallel/test_diffusers_adapters.py
@@ -114,7 +114,7 @@ class MochiPipelineTest(DiffusionPipelineTest):
 
     @pytest.mark.skipif("not torch.cuda.is_available()")
     @with_comms
-    @parametrize("dtype", [torch.bfloat16])
+    @parametrize("dtype", [torch.float16])
     @parametrize("device", ["cuda"])
     @parametrize(
         "parallelize,compile,use_ring",
